Move main view to diffrent package
