+++ app/src/main/AndroidManifest.xml
@@ -33,7 +33,7 @@
 			android:name="com.google.android.geo.API_KEY"
 			android:value="@string/google_maps_key" />
 
-		<activity android:name=".MainActivity">
+		<activity android:name=".mainView.MainActivity">
 			<intent-filter>
 				<action android:name="android.intent.action.MAIN" />
