+++ app/src/main/res/layout/fragment_map_view.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
 	xmlns:app="http://schemas.android.com/apk/res-auto"
 	xmlns:tools="http://schemas.android.com/tools"
-	tools:context=".MapViewFragment">
+	tools:context=".mainView.MapViewFragment">
 
 	<ScrollView
 		android:layout_width="match_parent"
