+++ app/src/main/res/navigation/navigation.xml
@@ -7,7 +7,7 @@
 
 	<fragment
 		android:id="@+id/mapViewFragment"
-		android:name="test.mug.espresso.MapViewFragment"
+		android:name="test.mug.espresso.mainView.MapViewFragment"
 		android:label="fragment_map_view"
 		tools:layout="@layout/fragment_map_view" />
 </navigation>
